+++ package.json
@@ -20,7 +20,6 @@
     "@gorhom/bottom-sheet": "^4",
     "@keystonehq/bc-ur-registry": "^0.5.2",
     "@ledgerhq/react-native-hw-transport-ble": "^6.27.8",
-    "@ngraveio/bc-ur": "^1.1.6",
     "@react-native-firebase/app": "^14.11.1",
     "@react-native-firebase/messaging": "^14.11.1",
     "@react-navigation/native": "^6.0.8",
