+++ src/hardware/index.ts
@@ -1,5 +1,6 @@
+import { Vault, VaultSigner } from 'src/core/wallets/interfaces/vault';
+
 import { SignerType } from 'src/core/wallets/enums';
-import { VaultSigner } from 'src/core/wallets/interfaces/vault';
 import WalletUtilities from 'src/core/wallets/operations/utils';
 import config from 'src/core/config';
 
@@ -79,4 +80,17 @@ export const getSignerNameFromType = (type: SignerType, isMock = false, isAmf = 
   } else {
     return name;
   }
+};
+
+export const getWalletConfig = ({ vault }: { vault: Vault }) => {
+  let line = '# Coldcard Multisig setup file (exported from Keeper)\n';
+  line += `Name: Keeper Vault\n`;
+  line += `Policy: ${vault.scheme.m} of ${vault.scheme.n}\n`;
+  line += `Format: P2SH-P2WSH\n`;
+  line += `\n`;
+  vault.signers.forEach((signer) => {
+    line += `Derivation: ${signer.xpubInfo.derivationPath}\n`;
+    line += `${signer.xpubInfo.xfp}: ${signer.xpub}\n\n`;
+  });
+  return line;
 };
